move fingerprint to seekbar package
